added landing page & redirect to main & about page
