+++ src/app/landing/landing.component.html
@@ -1,0 +1,2 @@
+<h1>Welcome to our app!</h1>
+<button routerLink="/main">Enter App</button>
